refactor : Changed datatype
Changed datatype of sum to long from double

+++ header/EvenFibo.h
@@ -8,7 +8,7 @@
 
 class EvenFibo {
 private:
-    double sum;
+    long sum;
 
 public:
